feat(db): 金額の処理をdoubleからdecimalにマイグレーション

+++ src/main/java/red/man10/man10bank/db/tables/AtmLog.kt
@@ -9,7 +9,7 @@ object AtmLog : Table<Nothing>("atm_log") {
     val id = int("id").primaryKey()
     val player = varchar("player")
     val uuid = varchar("uuid")
-    val amount = double("amount")
+    val amount = decimal("amount")
     val deposit = boolean("deposit")
     val date = datetime("date")
 }
