+++ src/main/java/red/man10/man10bank/db/tables/ChequeAndServerLoan.kt
@@ -9,7 +9,7 @@ object ChequeTbl : Table<Nothing>("cheque_tbl") {
     val id = int("id").primaryKey()
     val player = varchar("player")
     val uuid = varchar("uuid")
-    val amount = double("amount")
+    val amount = decimal("amount")
     val note = varchar("note")
     val date = datetime("date")
     val useDate = datetime("use_date")
@@ -23,8 +23,8 @@ object ServerLoanTbl : Table<Nothing>("server_loan_tbl") {
     val uuid = varchar("uuid")
     val borrowDate = datetime("borrow_date")
     val lastPayDate = datetime("last_pay_date")
-    val borrowAmount = double("borrow_amount")
-    val paymentAmount = double("payment_amount")
+    val borrowAmount = decimal("borrow_amount")
+    val paymentAmount = decimal("payment_amount")
     val failedPayment = int("failed_payment")
     val stopInterest = int("stop_interest") // TINYINT を int として扱う
 }
