+++ src/main/java/red/man10/man10bank/db/tables/UserBank.kt
@@ -1,9 +1,6 @@
 package red.man10.man10bank.db.tables
 
-import org.ktorm.schema.Table
-import org.ktorm.schema.double
-import org.ktorm.schema.int
-import org.ktorm.schema.varchar
+import org.ktorm.schema.*
 
 /**
  * user_bank テーブル
@@ -12,5 +9,5 @@ object UserBank : Table<Nothing>("user_bank") {
     val id = int("id").primaryKey()
     val player = varchar("player")
     val uuid = varchar("uuid")
-    val balance = double("balance")
+    val balance = decimal("balance")
 }
